Add Sub District and Remove Company Data

+++ customer/dto/address.go
@@ -5,7 +5,7 @@ type Address struct {
 	ID            uint   `json:"id"`
 	CustomerID    uint   `json:"customer_id"`
 	Line          string `json:"line"`
-	SubDistrictID uint   `json:"sub_district_id"`
+	SubDistrictID uint64 `json:"sub_district_id"`
 	DistrictID    uint   `json:"district_id"`
 	CityID        uint   `json:"city_id"`
 	ProvinceID    uint   `json:"province_id"`
