+++ customer/entity/address.go
@@ -5,7 +5,7 @@ type Address struct {
 	ID            uint   `gorm:"primaryKey;autoIncrement:true;Index"`
 	CustomerID    uint   `gorm:"type:INT;NOT NULL;Index"`
 	Line          string `gorm:"type:VARCHAR(255);"`
-	SubDistrictID uint   `gorm:"type:INT;NOT NULL;Index"`
+	SubDistrictID uint64 `gorm:"type:INT;NOT NULL;Index"`
 	DistrictID    uint   `gorm:"type:INT;NOT NULL;Index"`
 	CityID        uint   `gorm:"type:INT;NOT NULL;Index"`
 	ProvinceID    uint   `gorm:"type:INT;NOT NULL;Index"`
